Use images for ripple and background
Scaling a rounded circle-looking flex was not working on iOS.
The rounded corners would be lost during animation, giving the ripple
and the background an undesired squared aspect.
This change replaces the flex containers for both effects with an image
which can be exposed for the developer to choose different images to
change the color of the background and ripple.